View Changes. Fixed some bugs. Now >1 view change can happen. Tested with 7 view changes and 8 views

+++ coco/sign/constants.go
@@ -6,7 +6,7 @@ import "time"
 // TODO: can this be replaced by the application using the signer?
 var ROUND_TIME time.Duration = 10 * time.Second
 
-var RoundsPerView int = 3
+var RoundsPerView int64 = 3
 
 var FALSE int32 = 0
 var TRUE int32 = 1
